instDB: add robIdx and some TIME signals
