+++ src/main/scala/xiangshan/backend/rob/Rob.scala
@@ -96,24 +96,6 @@ class DebugLSIO(implicit p: Parameters) extends XSBundle {
   val debugLsInfo = Vec(exuParameters.LduCnt + exuParameters.StuCnt, Output(new DebugLsInfoBundle))
 }
 
-class InstInfoTable(implicit p: Parameters) extends XSBundle{
-  val globalID = UInt(XLEN.W)
-  val robIdx = UInt(log2Ceil(RobSize).W)
-  val instType = FuType()
-  val exceptType = ExceptionVec()
-  val ivaddr = UInt(VAddrBits.W)
-  val dvaddr = UInt(VAddrBits.W) // the l/s access address
-  val dpaddr = UInt(VAddrBits.W) // need the physical address when the TLB is valid
-  val tlbLatency = UInt(XLEN.W)  // original requirements is L1toL2TlbLatency
-  // val levelTlbHit = UInt(2.W) // 01, 10, 11(memory)
-  // val otherPerfNoteThing // FIXME: how much?
-  val accessLatency = UInt(XLEN.W)  // RS out time --> write back time
-  val executeLatency = UInt(XLEN.W)
-  val issueLatency = UInt(XLEN.W)
-  val lsInfo = new DebugLsInfo
-  val mdpInfo = new DebugMdpInfo
-}
-
 class RobPtr(implicit p: Parameters) extends CircularQueuePtr[RobPtr](
   p => p(XSCoreParamsKey).RobSize
 ) with HasCircularQueuePtrHelper {
@@ -1105,11 +1087,11 @@ class RobImp(outer: Rob)(implicit p: Parameters) extends LazyModuleImp(outer)
     val isWriteInstInfoTable = WireInit(Constantin.createRecord("isWriteInstInfoTable" + p(XSCoreParamsKey).HartId.toString))
     val instTableName = "InstTable" + p(XSCoreParamsKey).HartId.toString
     val instSiteName = "Rob" + p(XSCoreParamsKey).HartId.toString
-    val debug_instTable = ChiselDB.createTable(instTableName, new InstInfoTable)
+    val debug_instTable = ChiselDB.createTable(instTableName, new InstInfoEntry)
     // FIXME lyq: only get inst (alu, bj, ls) in exuWriteback
     for (wb <- exuWriteback) {
       when(wb.valid) {
-        val debug_instData = Wire(new InstInfoTable)
+        val debug_instData = Wire(new InstInfoEntry)
         val idx = wb.bits.uop.robIdx.value
         debug_instData.globalID := wb.bits.uop.ctrl.debug_globalID
         debug_instData.robIdx := idx
@@ -1121,10 +1103,12 @@ class RobImp(outer: Rob)(implicit p: Parameters) extends LazyModuleImp(outer)
         debug_instData.accessLatency := wb.bits.uop.debugInfo.writebackTime - wb.bits.uop.debugInfo.issueTime
         debug_instData.executeLatency := wb.bits.uop.debugInfo.writebackTime - wb.bits.uop.debugInfo.issueTime
         debug_instData.issueLatency := wb.bits.uop.debugInfo.issueTime - wb.bits.uop.debugInfo.selectTime
-        debug_instData.exceptType := wb.bits.uop.cf.exceptionVec
+        debug_instData.exceptType := Cat(wb.bits.uop.cf.exceptionVec)
         debug_instData.lsInfo := debug_lsInfo(idx)
         debug_instData.mdpInfo.ssid := wb.bits.uop.cf.ssid
         debug_instData.mdpInfo.waitAllStore := wb.bits.uop.cf.loadWaitStrict && wb.bits.uop.cf.loadWaitBit
+        debug_instData.issueTime := wb.bits.uop.debugInfo.issueTime
+        debug_instData.writebackTime := wb.bits.uop.debugInfo.writebackTime
         debug_instTable.log(
           data = debug_instData,
           en = wb.valid,
