+++ src/main/scala/xiangshan/package.scala
@@ -149,7 +149,8 @@ package object xiangshan {
   }
 
   object ExceptionVec {
-    def apply() = Vec(16, Bool())
+    val ExceptionVecSize = 16
+    def apply() = Vec(ExceptionVecSize, Bool())
   }
 
   object PMAMode {
